chore: bump minimum Deno version to 1.41.1 (#309)
This commit bumps the minimum required Deno version to 1.41.1 since in fact this
is the real minimum version required to actually use the current deployctl. The
following error log will appear if we try to run deployctl using Deno 1.41.0:

```console
# deno --version
deno 1.41.0 (release, aarch64-unknown-linux-gnu)
v8 12.1.285.27
typescript 5.3.3
# deno install -Arf jsr:@deno/deployctl
✅ Successfully installed deployctl
/root/.deno/bin/deployctl
# deployctl
A new release of Deno is available: 1.41.0 → 1.42.4 Run `deno upgrade` to install it.
error: Uncaught (in promise) TypeError: Importing a JSR package via an HTTPS URL is not implemented. Use a jsr: specifier instead for the time being.
  const fs = await import("./token_storage/fs.ts");
             ^
    at async https://jsr.io/@deno/deployctl/1.12.0/src/utils/token_storage.ts:33:14
```

Upgrading Deno to 1.41.1 solves this issue.

Also, this commit adds a clear statement to readme that 1.41.1+ is required to
use deployctl. This should help new users trying to use deployctl to some
extent, because it's very hard to figure out what they need to do to solve the
issue from the error message they run into when their Deno version is old.

Closes #308