feat(action): always print x-deno-ray for debuggability (#321)
This commit makes `x-deno-ray` values always printed in GitHub Actions as soon
as we get a response from the backend API server.

The primary motivation behind it is that it's hard to debug issues with
streaming APIs - in particular, `gitHubActionsDeploy`:

https://github.com/denoland/deployctl/blob/acb2ba33df6b146bc8705a1d1c16283d9fef9e78/src/utils/api.ts#L410-L424

This API returns 200 OK and then starts streaming the progress. But currently
`x-deno-ray` is not printed when the status code is 200. Problems that happen
during streaming, e.g. very slow file uploads, could be made easier to debug by
making `x-deno-ray` available in this case too.